% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_outside.R
\name{clip.outside}
\alias{clip.outside}
\title{Funcao para clipar um polígono pelo outro e incluir as partes dos poligonos que extrapolam a área clipada}
\usage{
clip.outside(area2clip, clippingArea)
}
\arguments{
\item{area2clip}{O spatialpolygonsdataframe a ser clipado}

\item{clippingArea}{O spatialpolygonsdataframe para ser usado como área para clipar}
}
\value{
A pasta cm todas as subpastas
}
\description{
Essa funcao cria um diretorio com nome dado e todas as subpastas.
}
\examples{
nome/Rscript
}
