% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clip}
\alias{clip}
\title{Funcao para clipar um polígono pelo outro}
\usage{
clip(area2clip, clippingArea)
}
\arguments{
\item{area2clip}{O spatialpolygonsdataframe a ser clipado}

\item{clippingArea}{O spatialpolygonsdataframe para ser usado como área para clipar}
}
\value{
A area clipada (spdf)
}
\description{
Essa função executa um clip de dois shps ou spatialpolygonsdataframe (spdf)
}
\examples{
clip<-function(area2clip, clippingArea)
}
