% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cria_repositorio.R
\name{Cria_repositorio}
\alias{Cria_repositorio}
\title{Funcao para criar um repositório padrao}
\usage{
Cria_repositorio(nome)
}
\arguments{
\item{nome}{O nome a ser dado para o projeto ou diretorio.}
}
\value{
A pasta cm todas as subpastas
}
\description{
Essa funcao cria um diretorio com nome dado e todas as subpastas.
}
\examples{
nome/Rscript
}
